Updated to send SMS
This version is updated to send info via SMS. Some modifications are required:
1)Change the trigger pin of ultrasonic_sensor_2 to another pin. It may be 36, 2 or any other not used already by te code
2)Add your code implemented to read the second ultrasonic sensor using the variable name "cant_latas_sonar_2". Note that the original "cant_latas_sonar" changed to "cant_latas_sonar_1"

Several changes were made:
a)"Serial" was modified to "SerialMon".  Nothing additional is required
b)A new variable "door_is" was added. Nothing additional is required

This code executes only once in order to avoid infinite SMS sendings. To re-execute the code just press the reset button in the TTGO board. Changes will be implemented after this version works perfectly, to send SMS only when the door is opened or by a timer (each hour, for instance)